Move *.ts files to src/ts and add scss plugin to rollup
