+++ package.json
@@ -6,13 +6,6 @@
 		"ts:check": "tsc --noEmit",
 		"ts:check-watch": "npm run ts:check -- --watch",
 		"ts:build": "tsc --emitDeclarationOnly",
-		"css:build": "npm run css:build-sass && npm run css:build-autoprefix && npm run css:build-cleancss",
-		"css:build-autoprefix": "postcss --use autoprefixer --map false --output dist/css/bin2dec.v$npm_package_version.css dist/css/bin2dec.v$npm_package_version.css",
-		"css:build-cleancss": "cleancss -o dist/css/bin2dec.v$npm_package_version.min.css dist/css/bin2dec.v$npm_package_version.css",
-		"css:build-sass": "sass --style expanded src/scss/global.scss dist/css/bin2dec.v$npm_package_version.css",
-		"css:clean": "rimraf css",
-		"css:deploy": "npm run css:clean && npm run css:build",
-		"css:watch": "npm run css:build-sass -- --watch",
 		"js:watch": "rimraf ./dev/dist && rollup -c rollup.config.dev.js -w",
 		"js:serve": "nodemon ./dev/serve.js",
 		"js:umd": "rimraf ./dist && rollup -c",
@@ -75,6 +68,7 @@
 		"prettier": "^2.6.2",
 		"rimraf": "^3.0.2",
 		"rollup": "^2.72.1",
+		"rollup-plugin-scss": "^3.0.0",
 		"rollup-plugin-uglify": "^6.0.4",
 		"sass": "^1.51.0",
 		"ts-jest": "^28.0.2",
