+++ src/ts/bin2dec.ts
