+++ tsconfig.json
@@ -13,7 +13,7 @@
 		"outDir": "./lib",
 		"paths": {
 			"@/*": [
-				"src/*",
+				"src/ts/*",
 			],
 		},
 		"lib": [
@@ -23,7 +23,7 @@
 		],
 	},
 	"include": [
-		"./src/**/*"
+		"./src/ts/**/*"
 	],
 	"exclude": [
 		"./node_modules",
